Calculate wavelet transformation

This function takes numeric signal data, and performs biwavelet transformation by calling biwavelet::wt(). It optionally returns a specific numeric matrix, for example when type="power.corr.norm" it returns the normalized power correlation matrix (see docs for biwavelet::wt() for more details).

calc_ephys_wavelet(x, iWt = NULL, return_type = c("wt", "m", "list"),
  type = c("power.corr.norm", "power", "power.corr", "power.norm",
  "wavelet", "phase"), x_condense_factor = 10, x_label_multiple = 5,
  i_step = 0.001, m = NULL, do.sig = FALSE, verbose = FALSE, ...)

Arguments

x
numeric matrix with two columns, where the first column contains neural signal data, and the second column contains timestamps in units of seconds. When iWt is supplied, x is not used.

iWt
list output from biwavelet::wt(), or NULL. When iWt is supplied, x is not used. When iWt is not supplied, it is only calculated when return_type is either "wt" or "list", or when "m" is not supplied.

return_type
character vector indicating the type of data to return: "wt" returns the full output of biwavelet::wt(); "m" returns the numeric frequency matrix; "list" returns a list including elements c("wt", "m") which include the previous two formats.

type
character indicating which frequency matrix to extract from the output of biwavelet::wt(). Note that when iWt is supplied, the matrix is taken directly from it. When m is supplied as a numeric matrix, it is used directly, and type is not used.

x_condense_factor
integer value indicating the level of compression to condense the numeric matrix, when return_type is either "m" or "list". When x_condense_factor=1 there is no compression, otherwise it calls condense_freq_matrix().

x_label_multiple
integer value indicating the multiple to use when defining labels to display. For example when x_label_multiple=5 labels are displayed at multiples of 5: 0, 5, 10, 15, 20, 25.

m
numeric matrix, optionally supplied as an alternative to extracting a numeric matrix from either iWt or the output of biwavelet::wt(x).

do.sig
logical indicating whether to calculate statistical significance. This argument is passed to biwavelet::wt().

verbose
logical indicating whether to print verbose output.

...
additional arguments are passed to biwavelet::wt() when iWt is not supplied.

Value

When return_type is "wt" the output of biwavelet::wt() is returned; when return_type is "m" a numeric matrix is returned, whose rownames are the "period" and colnames are the timestamp, this matrix may also be condensed when x_condense_factor is greater than 1; when return_type is "list" then a list is returned with elements "wt" and "m" containing the two previous return types. The numeric matrix "m" has usseful attributes: "lab_col" is a character vector of column labels at user-friendly positions; "lab_row" is a character vector of row labels at user-friendly positions, after converting "period" to frequency in Hertz (per second) units.

Since the data matrix is very large, and it is often not necessary to retain the full resolution of this data, it can be condensed by setting x_condense_factor to any value above 1. In that case condense_freq_matrix() is called to condense the numeric matrix to a smaller size before returning results. This process reduces data size substantially, while avoiding memory consumption of returning each numeric matrix produced by biwavelet::wt().

Details

This function attempts to return a numeric spectral frequency matrix, either by calling biwavelet::wt() then extracting the appropriate numeric matrix from the resulting list object; or by using the iWt data from a previous call to biwavelet::wt() to extract the appropriate numeric matrix; or by using a numeric matrix m supplied directly. In the latter case, biwavelet::wt() is not necessary and may be substantially faster.

See also